TaverbQuest_V1.4

+++ app/src/main/java/com/luminor/tavernquest/data/repository/RankingRepositoryImpl.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
 class RankingRepositoryImpl(private val dao: ActivityDao) : RankingRepository {
-    override fun observeTavern(tavernId: String, periodStart: Long?): Flow<List<RankingEntry>> = dao.observeTavernRanking(tavernId, periodStart ?: 0L).map { rows -> rows.map { RankingEntry(it.heroId, it.xp.coerceIn(0, Int.MAX_VALUE).toInt(), it.activeDays.coerceIn(0, Int.MAX_VALUE).toInt(), it.checkIns.coerceIn(0, Int.MAX_VALUE).toInt()) } }
+    override fun observeTavern(tavernId: String, periodStart: Long?): Flow<List<RankingEntry>> = dao.observeTavernRanking(tavernId, periodStart ?: 0L).map { rows -> rows.map { RankingEntry(it.heroId, it.xp.coerceIn(0L, Int.MAX_VALUE.toLong()).toInt(), it.activeDays.coerceIn(0L, Int.MAX_VALUE.toLong()).toInt(), it.checkIns.coerceIn(0L, Int.MAX_VALUE.toLong()).toInt()) } }
 }
